requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
celery==5.3.6
redis==5.0.1
python-decouple==3.8
psycopg2-binary
